requirements.txt
mido
python-rtmidi
sounddevice
numpy
pyserial
pygame
